requirements.txt
# Core Django and API Framework
django==4.2.6
djangorestframework==3.14.0
django-cors-headers==4.2.0
djangorestframework-simplejwt==5.3.0
# Data Manipulation and Preprocessing
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.1
openpyxl==3.1.2
scipy
statsmodels
# Database Drivers and ORM
SQLAlchemy==2.0.21
PyMySQL==1.1.0
psycopg2-binary==2.9.9
jsonfield
# Data Visualization
matplotlib==3.7.3
seaborn==0.13.0
plotly==5.17.0
xlrd

# Real-time Communication (WebSockets)
channels==4.0.0
# Utilities
python-dotenv==1.0.0
setuptools==68.2.2
# Development and Testing
pytest==7.4.2
pytest-django==4.7.0
black==23.9.1
flake8==6.1.0
twilio
cryptography

